support/mem: change type of __VFN_PAGESIZE to size_t
size_t is a more suitable type for something that is a size.

+++ include/vfn/support/mem.h
@@ -13,7 +13,7 @@
 #ifndef LIBVFN_SUPPORT_MEM_H
 #define LIBVFN_SUPPORT_MEM_H
 
-extern long __VFN_PAGESIZE;
+extern size_t __VFN_PAGESIZE;
 extern int __VFN_PAGESHIFT;
 
 void backtrace_abort(void);
